backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
pydantic_settings
python-dotenv
PyJWT
bcrypt
cryptography
python-multipart
pyyaml
langchain_community
langchain
dashscope
pymilvus
docx2txt
pypdf
mutagen
oss2
langchain_openai
anthropic
langchain_anthropic
matplotlib
openpyxl
tenacity